requirements.txt
# Core dependencies
python-dotenv>=1.0.0  
anthropic  
pdf2image>=1.16.3    